requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=-
#
aniso8601==9.0.1
    # via flask-restful
certifi==2022.6.15
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
cryptography==37.0.2
    # via -r requirements.in
flask==2.1.2
    # via
    #   -r requirements.in
    #   flask-restful
    #   flask-sqlalchemy
flask-restful==0.3.9
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via jinja2
pycparser==2.21
    # via cffi
pymysql==1.0.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via flask-restful
requests==2.28.0
    # via -r requirements.in
six==1.16.0
    # via flask-restful
sqlalchemy==1.4.37
    # via flask-sqlalchemy
urllib3==1.26.9
    # via requests
werkzeug==2.1.2
    # via flask
zipp==3.8.0
    # via importlib-metadata
